update test code

+++ tests/unittests/test-cclipboard.cpp
@@ -54,8 +54,8 @@ MATCHER_P(AnsiStringInGlobalMemory, expected_string, "") {
 MATCHER_P(SakuraFormatInGlobalMemory, expected_string, "") {
 	char* p = (char*)::GlobalLock(arg);
 	if (!p) return false;
-	int length = *(int*)p;
-	p += sizeof(int);
+	int length = *(size_t*)p;
+	p += sizeof(size_t);
 	std::wstring_view actual((const wchar_t*)p);
 	bool match = actual.size() == length && actual == expected_string;
 	::GlobalUnlock(arg);
@@ -228,7 +228,7 @@ protected:
 			std::wcscpy(p, unicodeText.data());
 		});
 		sakuraMemory.Lock<unsigned char>([=](unsigned char* p) {
-			*(int*)p = sakuraText.size();
+			*(size_t*)p = sakuraText.size();
 			std::wcscpy((wchar_t*)(p + sizeof(size_t)), sakuraText.data());
 		});
 		oemMemory.Lock<char>([=](char* p) {
